Make beam more flexible to be able to call it from C/C++ libraries
- move lua libs to their own namespace to prevent conflicts with other modules when doing 'require'
  This is a follow-up to pull-request #38 where 'require' to local lua libs were replaced by 'dofile' to make it work with lua interpreter.
  The downside of this first approch was that LUA_ENV was not used to find the libs, a precise path was needed instead.
  It worked well in command line but using this from a C lib was quite cumbersome as we needed to build absolute paths in the C/C++ code before calling lua.

  A better solution is to make good use of the namespaces.
  We can require our libs without worrying about naming conflicts and let the 'require' function find the files based on the lua env vars.

- beam.lua is now used as a module so it can be loaded and called from C/C++ libraries (it exposes methods to init and search as well as the parsed options)
  It's necessary to have separate functions to init and translate in a translation engine context where we want to load the model and other resources in memory once and then translate phrases over time.

- move runtime logic (reading from/writing to files, results printing, etc.) to new 'evaluate.lua' script.
  (README updated, just use 'th evaluate.lua...' instead of 'th beam.lua...')
