+++ s2sa/util.lua
